Add SQL migration and load-config func
This was adapted from maettu's code. Up migration
seems to work. Down needs some changes to work.

It stores a single number in a single row in a
migration table, and compares that number with the
highest integer in the checked-in SQL file.

In practical terms, we will likely be merging our
schema changes into the main schema file while we
are pre production.